+++ apps/docs/src/content/03-components/content/markdown/examples/headingOffset.tsx
@@ -1,0 +1,7 @@
+import { Markdown } from "@mittwald/flow-react-components";
+
+<Markdown headingOffset={1}>
+  {
+    "# Heading Level 1 -> 2 \n## Heading Level 2 -> 3 \n### Heading Level 3 -> 4 \n#### Heading Level 4 -> 5 \n##### Heading Level 5 -> 6 \n###### Heading Level 6 -> 6"
+  }
+</Markdown>;