+++ apps/docs/src/content/03-components/structure/list/examples/filterCategory.tsx
@@ -31,7 +31,9 @@ export default () => {
             : !verified;
         }}
       />
-      <DomainList.Item>
+      <DomainList.Item
+        textValue={(domain) => domain.domain}
+      >
         {(domain) => (
           <DomainList.ItemView>
             <Avatar