+++ apps/docs/src/content/03-components/structure/list/examples/filterGeneralCategory.tsx
@@ -41,7 +41,9 @@ export default () => {
         ]}
         matcher={() => true}
       />
-      <DomainList.Item>
+      <DomainList.Item
+        textValue={(domain) => domain.domain}
+      >
         {(domain) => (
           <DomainList.ItemView>
             <Avatar