+++ packages/components/src/integrations/react-hook-form/components/Field/Field.tsx
@@ -51,6 +51,7 @@ export function Field<T extends FieldValues>(props: FieldProps<T>) {
     SearchField: fieldProps,
     TextField: fieldProps,
     TextArea: fieldProps,
+    MarkdownEditor: fieldProps,
 
     Checkbox: {
       ...fieldProps,